requirements.txt
google-genai>=0.3.0
groq>=0.11.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
jinja2>=3.1.0
pillow>=10.0.0
faster-whisper>=1.0.0
pydantic>=2.6.0
